Allow arbitrary number of fread calls when receiving large data packets.
Prior code did a maxiumum of 40 read retries 5ms apart when reading a
large data packet. A message that requires more than 40 reads from
the socket will never complete!

This commit keeps reading from the socket as long as data is arriving
from the server within the configured default socket timeout. Error
handling could be improved, but this is the simplest patch I could come
up with to support the $verifyExactLength parameter as defined.